- added option --targetstudio which is used to specify a program   which creates the target device node to create the image on.   In contrast to the targetdevice option there will be no image   file dumped from the specified device. This option is mainly   used in SuSE Studio

+++ modules/KIWIImage.pm
@@ -603,6 +603,9 @@ sub createImageEXT {
 	if (! defined $name) {
 		return undef;
 	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
+	}
 	#==========================================
 	# Create filesystem on extend
 	#------------------------------------------
@@ -673,6 +676,9 @@ sub createImageReiserFS {
 	if (! defined $name) {
 		return undef;
 	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
+	}
 	#==========================================
 	# Create filesystem on extend
 	#------------------------------------------
@@ -734,6 +740,9 @@ sub createImageXFS {
 	my $name = $this -> preImage ($device);
 	if (! defined $name) {
 		return undef;
+	}
+	if ($this->{targetDevice}) {
+		$device = $this->{targetDevice};
 	}
 	#==========================================
 	# Create filesystem on extend
@@ -3025,14 +3034,30 @@ sub buildLogicalExtend {
 	#==========================================
 	# Create logical extend storage and FS
 	#------------------------------------------
-	unlink ($out);
-	my $data = qxx ("dd if=/dev/zero of=$out bs=1 seek=$seek count=1 2>&1");
-	my $code = $? >> 8;
-	if ($code != 0) {
-		$kiwi -> error  ("Couldn't create logical extend");
-		$kiwi -> failed ();
-		$kiwi -> error  ($data);
-		return undef;
+	if (($main::targetStudio) && (-x $main::targetStudio)) {
+		#==========================================
+		# Call custom image creation tool...
+		#------------------------------------------
+		my $data = qxx ("$main::targetStudio $seek 2>&1");
+		my $code = $? >> 8;
+		chomp $data;
+		if (($code != 0) || (! -b $data)) {
+			$kiwi -> error  ("Failed creating Studio storage device: $data");
+			$kiwi -> failed ();
+			return;
+		}
+		my $device = $data;
+		$this->{targetDevice} = $device;
+	} else {	
+		unlink ($out);
+		my $data = qxx ("dd if=/dev/zero of=$out bs=1 seek=$seek count=1 2>&1");
+		my $code = $? >> 8;
+		if ($code != 0) {
+			$kiwi -> error  ("Couldn't create logical extend");
+			$kiwi -> failed ();
+			$kiwi -> error  ($data);
+			return undef;
+		}
 	}
 	#==========================================
 	# Setup encoding
@@ -3136,7 +3161,7 @@ sub installLogicalExtend {
 	#==========================================
 	# dump image file from device if requested
 	#------------------------------------------
-	if ($device) {
+	if (($device) && (! $main::targetStudio)) {
 		$this -> cleanMount();
 		$name = $this -> buildImageName ();
 		my $dest = $this->{imageDest}."/".$name;
